pyproject.toml
[project]
name = "french-speech-to-text"
version = "0.1.0"
description = "French ASR experimentation pipeline comparing CTC vs LLM-adapter (SLAM-ASR architecture)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "transformers>=5.3.0",
    "datasets>=4.8.3",
    "wandb>=0.25.1",
    "hydra-core>=1.3.2",
    "jiwer>=4.0.0",
    "num2words>=0.5.14",
    "modal>=1.3.5",
    "torchcodec>=0.10.0",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
    "ty>=0.0.24",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
# E=pycodestyle, F=pyflakes, I=isort, UP=pyupgrade, B=bugbear, SIM=simplify
